DAOS-10244 test: Merge isolation test to agent_failure.py
Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: agent_failure
