--- src/tests/ftest/deployment/agent_failure_isolate.yaml
@@ -1,69 +1,0 @@
-hosts:
-    test_servers:
-        - server-A
-        - server-B
-        - server-C
-        - server-D
-        - server-E
-        - server-F
-    test_clients:
-        - client-A
-        - client-B
-
-timeout: 300
-
-setup:
-    start_agents_once: False
-
-daos_server:
-  pattern_timeout: 60
-
-server_config:
-    name: daos_server
-    engines_per_host: 2
-    servers:
-        0:
-            pinned_numa_node: 0
-            nr_xs_helpers: 1
-            fabric_iface: ib0
-            fabric_iface_port: 31317
-            log_file: daos_server0.log
-            bdev_class: nvme
-            bdev_list: ["0000:00:00.0"]
-            scm_class: dcpm
-            scm_list: ["/dev/pmem0"]
-            scm_mount: /mnt/daos0
-        1:
-            pinned_numa_node: 1
-            nr_xs_helpers: 1
-            fabric_iface: ib1
-            fabric_iface_port: 31417
-            log_file: daos_server1.log
-            bdev_class: nvme
-            bdev_list: ["0000:00:00.1"]
-            scm_class: dcpm
-            scm_list: ["/dev/pmem1"]
-            scm_mount: /mnt/daos1
-
-pool:
-    size: 100%
-    control_method: dmg
-
-container:
-    type: POSIX
-    control_method: daos
-
-ior:
-    client_processes:
-        ppn: 10
-    iorflags:
-        flags: "-k -v -w -W"
-        api: DFS
-        # This test doesn't work if the IOR runs too fast. In that case, increase
-        # block_size and/or decrease transfer_size. To decrease transfer_size, divide it
-        # by 2, 4, 8, etc.
-        transfer_size: '256K'
-        block_size: '5G'
-        sw_deadline: 60
-        dfs_oclass: "SX"
-        dfs_dir_oclass: "SX"
